Separate execution class and move to execution/oneshot
